fix: missing types in UserConfig
