+++ .changeset/eleven-tips-try.md
@@ -1,0 +1,5 @@
+---
+'@modern-js/plugin-tailwindcss': patch
+---
+
+fix: UserConfig is missing tools.tailwindcss type
