+++ .changeset/kind-donuts-teach.md
@@ -1,0 +1,5 @@
+---
+'@modern-js/plugin-sass': patch
+---
+
+fix: UserConfig is missing tools.sass type
