+++ .changeset/real-phones-fold.md
@@ -1,0 +1,5 @@
+---
+'@modern-js/plugin-less': patch
+---
+
+fix: UserConfig is missing tools.less type
