+++ .changeset/young-dots-turn.md
@@ -1,0 +1,5 @@
+---
+'@modern-js/plugin-bff': patch
+---
+
+fix: allow reference typings to index file
