+++ packages/cli/plugin-bff/src/index.ts
@@ -1,1 +1,3 @@
+import './types';
+
 export * from './constants';
