+++ packages/cli/plugin-bff/src/types.ts
@@ -1,0 +1,12 @@
+import '@modern-js/core';
+
+declare module '@modern-js/core' {
+  interface UserConfig {
+    bff?: {
+      prefix?: string;
+      requestCreator?: string;
+      fetcher?: string;
+      proxy: Record<string, any>;
+    };
+  }
+}
