+++ packages/cli/plugin-sass/src/index.ts
@@ -1,1 +1,3 @@
+import './types';
+
 export { default } from './cli';
