+++ packages/cli/plugin-sass/src/types.ts
@@ -1,9 +1,11 @@
-import { Options } from 'sass';
+import type { Options } from 'sass';
 import '@modern-js/core';
+
+type SassOptions = Options<'sync'> | Options<'async'>;
 
 declare module '@modern-js/core' {
   interface ToolsConfig {
     // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-    sass?: Options | ((options: Options) => Options | void);
+    sass?: SassOptions | ((options: SassOptions) => SassOptions | void);
   }
 }
